pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel>=0.41.0"]
build-backend = "setuptools.build_meta"

[project]
name = "retrieva"
version = "0.1.0"
description = "A configurable Retrieval-Augmented Generation (RAG) platform"
readme = "docs/README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Retrieva Team" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Web framework
    "fastapi>=0.110.0,<1.0.0",
    "uvicorn[standard]>=0.27.0,<1.0.0",
    "python-multipart>=0.0.9,<1.0.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.25,<3.0.0",
    "asyncpg>=0.29.0,<1.0.0",
    "psycopg2-binary>=2.9.9,<3.0.0",
    "alembic>=1.13.0,<2.0.0",

    # Configuration & validation
    "pydantic[email]>=2.5.0,<3.0.0",
    "pydantic-settings>=2.1.0,<3.0.0",
    "pyyaml>=6.0.1,<7.0.0",

    # Task queue
    "celery[redis]>=5.3.0,<6.0.0",
    "redis>=5.0.0,<6.0.0",

    # Vector database
    "qdrant-client>=1.7.0,<2.0.0",

    # LLM providers
    "openai>=1.12.0,<2.0.0",
    "anthropic>=0.18.0,<1.0.0",

    # Embeddings
    "sentence-transformers>=2.3.0,<3.0.0",

    # Web crawling & scraping
    "httpx>=0.27.0,<1.0.0",
    "beautifulsoup4>=4.12.0,<5.0.0",
    "trafilatura>=1.6.0,<2.0.0",

    # File extractors
    "pypdf2>=3.0.0,<4.0.0",
    "python-docx>=1.1.0,<2.0.0",
    "openpyxl>=3.1.0,<4.0.0",
    "pandas>=2.1.0,<3.0.0",

    # Keyword search
    "rank-bm25>=0.2.2,<1.0.0",

    # Authentication
    "bcrypt>=4.0.0,<5.0.0",
    "python-jose[cryptography]>=3.3.0,<4.0.0",

    # Cloud storage
    "boto3>=1.34.0,<2.0.0",

    # Rate limiting
    "slowapi>=0.1.9,<1.0.0",

    # CLI
    "click>=8.1.0,<9.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]
test = [
    "pytest>=8.0.0,<9.0.0",
    "pytest-asyncio>=0.23.0,<1.0.0",
    "pytest-cov>=4.1.0,<6.0.0",
    "httpx>=0.27.0,<1.0.0",
    "factory-boy>=3.3.0,<4.0.0",
    "aiosqlite>=0.19.0,<1.0.0",
]

[project.scripts]
retrieva = "core.cli:main"

[project.entry-points."retrieva.plugins"]

[tool.setuptools.packages.find]
include = ["api*", "core*", "workers*", "plugins*"]

# ---------- Tool configuration ----------

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "S",   # flake8-bandit
    "T20", # flake8-print
    "SIM", # flake8-simplify
]
ignore = ["S101", "S106"]

[tool.ruff.lint.isort]
known-first-party = ["api", "core", "workers", "plugins"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = [
    "--strict-markers",
    "--tb=short",
    "-q",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests (require running services)",
    "e2e: End-to-end tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["api", "core", "workers"]
branch = true
omit = [
    "*/tests/*",
    "*/alembic/*",
]

[tool.coverage.report]
fail_under = 80
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.",
]
